requirements.txt
Django>=3.0.8,<3.1.0
djangorestframework>=3.11.0,<3.12.0
flake8>=3.8.3,<3.9.0
django-mysql>=2.2.0,<2.3.0
django-phone-field>=1.8.1,<1.9.0